Fixed bug in FindUniqueWords that was adding extra whitespace
The two removed regexes were matching on any whitespace (including none)
and adding a single space. I think this was intended to collapse
consecutive whitespace into a single space, but it was adding an extra
space to the end of any string that passed through this section of code.
Further, there was another regex shortly before it that removed all
whitespace entirely, so it was doing nothing other than adding a single
space to the end of every string.