Adjust BackgroundProcess reports to contain more info
Posda::BackgroundProcess->CreateReport now returns a
Posda::BackgroundProcess::Report object. This object has a print()
method which allowes backward compatiblity with the previous
implementation. When stringified, it becomes the URL.

It is now possible to access a number of details about the underlying
report, such as: file_id, downloadable_file_id, path, url. The close()
method must be called before these details become available.

If you manually call close on a report, it will be up to you to record
the URL (or other info) in the script email. If you do not call close,
the previous behavior is kept: the report URL will automatically be
added to the email at the end.

This resolves PT-494.