Add filter to Table objects
Also create the DbIf::Table object to replace the simple
hash that was being used. Filtering logic was added to this
new class.

This will need some extra work, as the interface is a bit
buggy (causing the user's focus to be lost when typing in the
filter boxes), but it is fully functional.

This resolves PT-528