Add collection and site back to ProduceInitialAnonymizerCommandsTp.pl
A mistake was made previously, and the collection and site
parameters were removed from this script. However, as this script will
be run on a timepoint filled with files that lack Group 13 (CTP)
tags, it will not actually be possible to determine the correct
collection and site without having them passed as parameters. This
commit fixes that.

This relates to PT-717