Adjust BackgroundPrivateDispositionsTp*.pl to set Manual Update flag
This commit adds a new method to BackgroundProcess to allow a script
to set the manual update flag without having to call Finish. It then
adjusts the two BackgroundPrivateDispositionsTp* scripts to call this
method right before processing begins.

This commit also supplies an example query in the API for getting a list
of Public Send statuses that are ready to be tracked.