Use collection/site from file when sending to public
This commit modifies ApplyPrivateDispositionUnconditionalDate2.pl to use
the Collection Name and Site Name contained in the Group 13 tags
in the file being processed, rather than expecting those values to
be passed in on the command line.

This fixes an issue where having a timepoint with multiple collections
and sending it to public would place all of the files in a single
collection.