Add Third Party Analysis URL support to send to NIBA
This commit makes the required changes to add the Third Party Analysis
(TPA) URL when sending to NBIA/Public.

If a TPA URL is set for an activity, when that acitivty is sent to
public, the TPA URL will be sent along. If no URL is set, nothing extra
is sent during the NBIA submission.

This relates to PT-788 and PT-790