make BackgroundPrivateDispositionsTp[Baseline].tp send timepoint as batch
This commit modifies the two BackgroundPrivateDispositionsTp scripts so
that they send the Activity Timepoint ID as the Batch when sending to
NBIA. This will allow easy backwards tracking of files in NBIA to the
exact activity timepoint they came from.

This commit also fixes a small bug with these scripts that would cause
them to break when called with empty parameters.

This resolves PT-941