Increase the wait time between loops for refilling the FFP queue
This is an attempt to resolve a problem when importing large files:
If a large file is the only file in the import queue, and it takes more
than the timeout time (originally 1 second) to process, it would get
picked up and added to the queue a second time while a worker was in
the middle of processing it. This leads to the file getting processed
mutliple times.

This is a bandaid fix, but may resolve it for now.