Add parameter for node_hostname to the worker/status/*/running endpoint
This was always intended to be here, but was mistakenly omitted. This
records which worker node actually took the work item and completed it.
While running in docker-compose, the hostnames are random strings of
hex; however when running on dedicated hosts or on k8s, they should be
useful values.