Adjust magicka to use one invocation per frame when rendering series
The previous attempt to fix tempfile/memory usage was insufficient;
it appears that ImageMagick loads the entire imageset into memory when
doing a "filelist" conversion, even if it writes multiple output files.

This is a bandaid fix that switches to an IM invocation for each frame
in an IEC during rendering. It is much slower than previously, but
appears to work for all series types.