Fix bugs in default nginx config
The location entires in the nginx config were incorrect, and it was
causing them to match on some URLs they shouldn't, especially the
kaleidoscope entry. This commit adjusts all location directives to match
trailing slashes, which should fix this issue.

There is, however, a small chance it could introduce bugs in some of the
application endpoints (like /viewer/ or /k/, etc).