Add runtime metrics collection to background processes
This change uses the unix time utility to collect runtime statistics
about every background process. It includes memory and CPU usage,
and some other values. Changes are made to the API to pass these
through, and a new column metrics is added to the work table to track
them.

This relates to PT-1011