Add a check to skip multiframe objects
The current method of rendering projections doesn't work for multiframe
files, and sometimes can render just the first frame. This change makes
it always abort for any IEC that contains a multiframe file. This is not
a long-term solution but will help avoid falsely flagging IECs as
PHI-free.

This may resolve PT-1148