Add overwrite configuration option to Ream
REAM_OVERWRITE can now be set/unset in the ream configuration
(ream.env) to control whether NBIA is asked to overwrite files
when submitting.

Note that this is the default behavior of NBIA if the flag is unset
during the request, so this new default value doesn't change that
behavior.