Fix mistake in GetBasicFileInfo, greatly speeding it up
This query was joining against activity_timepoint_file for some reason,
but was not actually using that table for anything. Perhaps there was a
planned return column that never got added. There is no index on
activity_timepoint_file.file_id, which was the source of the slowness.
This commit just removes the join against the table altogether, as it
was unused.