Adjust posda-api so it handles SIGTERM and stops quickly
The bash script that starts uvicorn/fastapi had to be updated to use
exec, so that uvicorn becomes pid 1 inside the container.

The FastAPI main script also had to be adjusted to handle SIGTERM. At
this time, it just immediately exists, without doing any kind of cleanup.
This should be improved in the future.