Add error message to Quince for multiframe files
Due to other fixes in the db, Quince will now display the first frame of
multiframe files, but gives no indication that it is a multiframe, which
could lead to curators missing the other frames. This temporary fix just
disables the display of multiframe files entirely, instead displaying an
error message (Multiframe File; Use other viewer).