Make some changes necessary to support Cornerstone v2.0 Nifti
Cornerstone 2.0 made some major changes in how Nifti files were loaded
and processed, and one of those is that it determines if the file is
compressed by looking at the end of the url (.gz vs not). To support
proper handling of these files, this commit adds an alias to the
files/data endpoint of files/data.gz. This does nothing other than trick
the Cornerstone code into believing the file is compressed.

This commit also adds an is_zipped field to the return of the nifti
endpoint.