Introduce NewApplyMasks, a complete rewrite of ApplyMasks
This version handles orphans and referencing files, and does
almost all processing itself. It does still generate an edit skeleton
for moving the pre-masked files, as that operation needs to have an
email to link to (so this script cannot create a new timepoint in
a second activity without causing issues).