Add new /activities endpoint for UI operations involving Activities
This adds a set of endpoints under /v1/activities that will be useful
for new and exsiting UI code. This includings being able to list
all activities, list timeline entries for an activity, list details
about a specific timeline entry, and retrieve the input and output data
for a given timeline entry / subprocess.