Fix bug in NewTry for parsing of output from file
NewTry calls file to determine the file type of files it has
determined are not DICOM. It uses a regex to break the result
on the first colon, but some nifti files return a file type that
contains more than one colon. Adjusting this regex to be "non-greedy"
fixes the issue so the correct type information is returned.