Add new query ResetExportEvent and supporting code
This new query allows a normal user to reset an export_event
by supplying the export_event_id as well as the expected number
of files that need to be reset.

This commit also adds support for a '-- select' prefix to a query
to force it to be treated as a select query (even if it's an update,
or it starts with a CTE).